[core] Make useForkRef variadic (#27939)
