+++ packages/mui-base/src/SelectUnstyled/SelectUnstyled.tsx
@@ -119,7 +119,7 @@ const SelectUnstyled = React.forwardRef(function SelectUnstyled<TValue extends {
     setButtonDefined(element != null);
   }, []);
 
-  const handleButtonRef = useForkRef(forwardedRef, useForkRef(buttonRef, handleButtonRefChange));
+  const handleButtonRef = useForkRef(forwardedRef, buttonRef, handleButtonRefChange);
 
   React.useEffect(() => {
     if (autoFocus) {
