+++ packages/mui-base/src/utils/useSlotProps.ts
@@ -80,7 +80,8 @@ export default function useSlotProps<
 
   const ref = useForkRef(
     internalRef,
-    useForkRef(resolvedComponentsProps?.ref, parameters.additionalProps?.ref),
+    resolvedComponentsProps?.ref,
+    parameters.additionalProps?.ref,
   ) as ((instance: any | null) => void) | null;
 
   const props: UseSlotPropsResult<ElementType, SlotProps, AdditionalProps, OwnerState> =
