+++ packages/mui-material/src/InputBase/InputBase.js
@@ -300,9 +300,13 @@ const InputBase = React.forwardRef(function InputBase(inProps, ref) {
       }
     }
   }, []);
-  const handleInputPropsRefProp = useForkRef(inputPropsProp.ref, handleInputRefWarning);
-  const handleInputRefProp = useForkRef(inputRefProp, handleInputPropsRefProp);
-  const handleInputRef = useForkRef(inputRef, handleInputRefProp);
+
+  const handleInputRef = useForkRef(
+    inputRef,
+    inputRefProp,
+    inputPropsProp.ref,
+    handleInputRefWarning,
+  );
 
   const [focused, setFocused] = React.useState(false);
   const muiFormControl = useFormControl();
